requirements.txt
Flask==1.1.2
Flask-Dropzone==1.5.4
Flask-Uploads==0.2.1
gunicorn==20.0.4
python-decouple==3.3
Cython==0.29.20
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
matplotlib==3.2.2
mrcnn==0.2
numpy==1.16.4
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
Pillow==7.2.0
pycocotools==2.0.1
scikit-image==0.17.2
scipy==1.5.0
tensorboard==1.9.0
tensorflow==1.9.0